requirements.txt
absl-py==0.13.0
altair==4.1.0
argon2-cffi==21.1.0
astor==0.8.1
astunparse==1.6.3
attrs==21.2.0
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
backports.zoneinfo==0.2.1
base58==2.1.1
bleach==4.1.0
blinker==1.4
cachetools==4.2.4
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.7
clang==5.0
click==7.1.2
cloudpickle @ file:///Users/ktietz/demo/mc3/conda-bld/cloudpickle_1629142150447/work
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
cycler==0.10.0
cytoolz==0.11.0
dask @ file:///tmp/build/80754af9/dask-core_1630523592567/work
debugpy @ file:///C:/ci/debugpy_1629222854708/work
decorator @ file:///tmp/build/80754af9/decorator_1621259047763/work
defusedxml==0.7.1
docutils==0.17.1
entrypoints==0.3
flatbuffers==1.12
fonttools==4.25.0
fsspec @ file:///tmp/build/80754af9/fsspec_1626383727127/work
gast==0.4.0
gitdb==4.0.9
GitPython==3.1.24
google-auth-oauthlib==0.4.5
google-pasta==0.2.0
grpcio==1.39.0
h5py==3.1.0
idna==3.3
imagecodecs @ file:///C:/ci/imagecodecs_1623228459942/work
imageio @ file:///tmp/build/80754af9/imageio_1617700267927/work
imgaug @ file:///tmp/build/80754af9/imgaug_1605226543559/work
intel-openmp==2021.3.0
ipykernel @ file:///C:/ci/ipykernel_1629995240690/work/dist/ipykernel-6.2.0-py3-none-any.whl
ipython==7.27.0
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
ipywidgets==7.6.5
jedi @ file:///C:/ci/jedi_1611315808330/work
Jinja2==3.0.2
joblib==1.0.1
jsonschema==4.1.2
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1616770841739/work
jupyter-core @ file:///C:/ci/jupyter_core_1612213356021/work
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
locket==0.2.1
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline @ file:///tmp/build/80754af9/matplotlib-inline_1628242447089/work
memory-profiler==0.58.0
MiniSom==2.2.9
mistune==0.8.4
mkl==2021.3.0
mkl-fft==1.3.0
mkl-random @ file:///C:/ci/mkl_random_1626186184278/work
mkl-service==2.4.0
munkres==1.1.4
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx @ file:///tmp/build/80754af9/networkx_1627459939258/work
notebook==6.4.5
numpy==1.21.2
olefile @ file:///Users/ktietz/demo/mc3/conda-bld/olefile_1629805411829/work
opencv-contrib-python==4.4.0.46
opencv-python==4.5.3.56
opt-einsum==3.3.0
packaging @ file:///tmp/build/80754af9/packaging_1625611678980/work
pandas==1.3.4
pandocfilters==1.5.0
parso @ file:///tmp/build/80754af9/parso_1617223946239/work
partd @ file:///tmp/build/80754af9/partd_1618000087440/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow @ file:///C:/ci/pillow_1625663293593/work
prometheus-client==0.12.0
prompt-toolkit==3.0.20
protobuf==3.19.1
psutil==5.8.0
pyarrow==6.0.0
pycparser==2.20
pydeck==0.7.1
Pygments @ file:///tmp/build/80754af9/pygments_1629234116488/work
pyparsing @ file:///home/linux1/recipes/ci/pyparsing_1610983426697/work
pyrsistent==0.18.0
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets @ file:///C:/ci/pywavelets_1601658407916/work
pywin32==228
pywinpty==1.1.5
PyYAML==5.4.1
pyzmq @ file:///C:/ci/pyzmq_1628276144921/work
requests==2.26.0
scikit-image==0.18.3
scikit-learn==0.24.2
scipy @ file:///C:/ci/scipy_1630606942584/work
Send2Trash==1.8.0
Shapely==1.7.1
six @ file:///tmp/build/80754af9/six_1623709665295/work
smmap==5.0.0
smote-variants==0.4.0
statistics==1.0.3.5
streamlit==1.1.0
streamlit-player==0.1.5
tensorboard==2.6.0
tensorflow-estimator==2.6.0
tensorflow-gpu==2.6.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==2.2.0
tifffile==2021.8.8
toml==0.10.2
toolz @ file:///home/linux1/recipes/ci/toolz_1610987900194/work
torchaudio==0.9.0
torchvision==0.10.0+cu111
tornado @ file:///C:/ci/tornado_1606942392901/work
tqdm @ file:///tmp/build/80754af9/tqdm_1629302309755/work
traitlets @ file:///home/ktietz/src/ci/traitlets_1611929699868/work
typing-extensions==3.7.4.3
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
validators==0.18.2
watchdog==2.1.6
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
widgetsnbextension==3.5.2
wincertstore==0.2
wrapt==1.12.1
xgboost==1.4.2
